Re-namespace statement objects
