+++ src/Line/Statement/AssignmentStatementInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace webignition\BasilCompilableSource\Line;
+namespace webignition\BasilCompilableSource\Line\Statement;
 
 use webignition\BasilCompilableSource\VariablePlaceholder;
