+++ src/Line/Statement/StatementInterface.php
@@ -2,7 +2,9 @@
 
 declare(strict_types=1);
 
-namespace webignition\BasilCompilableSource\Line;
+namespace webignition\BasilCompilableSource\Line\Statement;
+
+use webignition\BasilCompilableSource\Line\ExpressionInterface;
 
 interface StatementInterface extends ExpressionInterface
 {
